Cleanup panache generation
